refactor(#933): extract auth command runtime into tau-provider
