--- crates/tau-coding-agent/src/cli_executable.rs
@@ -1,1 +1,0 @@
-pub(crate) use tau_provider::is_executable_available;
