+++ crates/tau-provider/Cargo.toml
@@ -18,6 +18,7 @@ tau-cli = { path = "../tau-cli" }
 tau-core = { path = "../tau-core" }
 tokio.workspace = true
 tracing.workspace = true
+wait-timeout.workspace = true
 
 [dev-dependencies]
 httpmock = "0.8"
